+++ src/v3/V3Response.cpp
@@ -61,7 +61,7 @@ bool etcdv3::V3Response::has_values() const
   return values.size() > 0;
 }
 
-int etcdv3::V3Response::get_compact_revision() const
+int64_t etcdv3::V3Response::get_compact_revision() const
 {
   return compact_revision;
 }
